Fold restclient.IntegrationTest into testutils
IntegrationTest now provide OAuth and CourtCase mocks
